+++ DeeployTest/testMVP.py
@@ -26,7 +26,7 @@
 import os
 import sys
 from collections import OrderedDict
-from typing import List
+from typing import List, Tuple
 
 import numpy as np
 import onnx
@@ -76,7 +76,7 @@ def _filterSchedule(schedule: List[List[gs.Node]], layerBinding: 'OrderedDict[st
 
 
 def setupDeployer(graph: gs.Graph, memoryHierarchy: MemoryHierarchy, defaultTargetMemoryLevel: MemoryLevel,
-                  defaultIoMemoryLevel: MemoryLevel, verbose: CodeGenVerbosity) -> NetworkDeployer:
+                  defaultIoMemoryLevel: MemoryLevel, verbose: CodeGenVerbosity) -> Tuple[NetworkDeployer, bool]:
 
     inputTypes = {}
     inputOffsets = {}
@@ -143,7 +143,7 @@ def setupDeployer(graph: gs.Graph, memoryHierarchy: MemoryHierarchy, defaultTarg
     deployer.tiler.memoryAllocStrategy = args.memAllocStrategy
     deployer.tiler.searchStrategy = args.searchStrategy
 
-    return deployer
+    return deployer, signProp
 
 
 if __name__ == '__main__':
@@ -263,14 +263,13 @@ if __name__ == '__main__':
     memoryHierarchy = MemoryHierarchy(memoryLevels)
     memoryHierarchy.setDefaultMemoryLevel(args.defaultMemLevel)
 
-    deployer = setupDeployer(graph,
-                             memoryHierarchy,
-                             defaultTargetMemoryLevel = L1,
-                             defaultIoMemoryLevel = memoryHierarchy.memoryLevels[args.defaultMemLevel],
-                             verbose = verbosityCfg)
+    deployer, signProp = setupDeployer(graph,
+                                       memoryHierarchy,
+                                       defaultTargetMemoryLevel = L1,
+                                       defaultIoMemoryLevel = memoryHierarchy.memoryLevels[args.defaultMemLevel],
+                                       verbose = verbosityCfg)
 
     platform = deployer.Platform
-    signProp = False
 
     for index, num in enumerate(test_inputs):
         _type, offset = inferTypeAndOffset(num, signProp)