+++ DeeployTest/testUtils/codeGenerate.py
@@ -30,6 +30,7 @@ import numpy as np
 
 from Deeploy.DeeployTypes import CodeGenVerbosity, ConstantBuffer, NetworkDeployer, VariableBuffer
 from Deeploy.Targets.MemPool.Platform import MemPoolPlatform
+from Deeploy.Targets.PULPOpen.Platform import MemoryPULPPlatform, MemoryPULPPlatformWrapper, PULPPlatform
 
 _TEXT_ALIGN = 30
 
@@ -81,12 +82,11 @@ def generateTestInputsHeader(deployer: NetworkDeployer, test_inputs: List) -> st
             list_str = (", ").join([str(x) for x in values])
 
         # WIESEP: Arrays have to be 4 byte aligned (at least in banshee)
-        bytes = (len(values) * typeWidth) // 8
-        if bytes % 4 != 0:
-            paddingBytes = bytes % 4
-            paddingElements = paddingBytes * 8 // typeWidth
-            list_str += ", "
-            list_str += (", ").join([str(0) for _ in range(paddingElements)])
+        total_bytes = (values.size * typeWidth) // 8
+        pad_bytes = (-total_bytes) % 4
+        if pad_bytes:
+            paddingElements = (pad_bytes * 8 + typeWidth - 1) // typeWidth
+            list_str += ", " + (", ").join("0" for _ in range(paddingElements))
 
         retStr += list_str
         retStr += "};\n"
@@ -116,12 +116,11 @@ def generateTestOutputsHeader(deployer: NetworkDeployer, test_outputs: List[np.n
             list_str = (", ").join([str(x) for x in values])
 
         # WIESEP: Arrays have to be 4 byte aligned (at least in banshee)
-        bytes = (len(values) * typeWidth) // 8
-        if bytes % 4 != 0:
-            paddingBytes = bytes % 4
-            paddingElements = paddingBytes * 8 // typeWidth
-            list_str += ", "
-            list_str += (", ").join([str(0) for _ in range(paddingElements)])
+        total_bytes = (len(values) * typeWidth) // 8
+        pad_bytes = (-total_bytes) % 4
+        if pad_bytes:
+            paddingElements = (pad_bytes * 8 + typeWidth - 1) // typeWidth
+            list_str += ", " + (", ").join("0" for _ in range(paddingElements))
 
         retStr += list_str
         retStr += "};\n"
@@ -144,11 +143,18 @@ def generateTestNetworkHeader(deployer: NetworkDeployer) -> str:
     #include <stdlib.h>
     """
     retStr += deployer.generateIncludeString()
-    retStr += """
-    void RunNetwork(uint32_t core_id, uint32_t numThreads);
-    void InitNetwork(uint32_t core_id, uint32_t numThread);
+    if isinstance(deployer.Platform, (PULPPlatform, MemoryPULPPlatform, MemoryPULPPlatformWrapper)):
+        retStr += """
+        void RunNetwork();
+        void InitNetwork();
 
-    """
+        """
+    else:
+        retStr += """
+        void RunNetwork(uint32_t core_id, uint32_t numThreads);
+        void InitNetwork(uint32_t core_id, uint32_t numThread);
+
+        """
 
     retStr += deployer.generateIOBufferInitializationCode()
     retStr += """
@@ -181,6 +187,11 @@ def generateTestNetworkImplementation(deployer: NetworkDeployer, verbosityCfg: C
         retStr += """
         void RunNetwork(__attribute__((unused)) uint32_t core_id, __attribute__((unused)) uint32_t numThreads){
         """
+    elif isinstance(deployer.Platform, (PULPPlatform, MemoryPULPPlatform, MemoryPULPPlatformWrapper)):
+        retStr += """
+        void RunNetwork(){
+        """
+        retStr += deployer.generateInferenceInitializationCode()
     else:
         retStr += """
         void RunNetwork(__attribute__((unused)) uint32_t core_id, __attribute__((unused)) uint32_t numThreads){
@@ -188,11 +199,18 @@ def generateTestNetworkImplementation(deployer: NetworkDeployer, verbosityCfg: C
         retStr += deployer.generateInferenceInitializationCode()
 
     retStr += deployer.generateFunction(verbosityCfg)
-    retStr += """
-    }
+    if isinstance(deployer.Platform, (PULPPlatform, MemoryPULPPlatform, MemoryPULPPlatformWrapper)):
+        retStr += """
+        }
 
-    void InitNetwork(__attribute__((unused)) uint32_t core_id, __attribute__((unused)) uint32_t numThreads){
-    """
+        void InitNetwork(){
+        """
+    else:
+        retStr += """
+        }
+
+        void InitNetwork(__attribute__((unused)) uint32_t core_id, __attribute__((unused)) uint32_t numThreads){
+        """
     retStr += deployer.generateEngineInitializationCode()
     retStr += deployer.generateBufferAllocationCode()
     retStr += """