+++ TargetLibraries/PULPOpen/src/RQiHardswish.c
@@ -37,7 +37,7 @@ void RQiHardswish_s8_s8_plp(int8_t *input, int8_t *output, int32_t size,
   rnd = (1 << (shift - 1));
 
   int8_t core_id = pi_core_id();
-  int8_t log2Core = log2(NUM_CORES);
+  int8_t log2Core = LOG2(NUM_CORES);
   int16_t chunk = (size >> log2Core) + ((size & (NUM_CORES - 1)) != 0);
   int16_t chunk_start = MIN(chunk * core_id, size);
   int16_t chunk_stop = MIN(chunk_start + chunk, size + 1);